Final parte 4

+++ part4/bloglist/controllers/blogs.js
@@ -1,25 +1,36 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
-  response.json(blogs)
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
+  response.json(blogs.map(blog => blog.toJSON()))
 })
 
-blogsRouter.get('/:id', (request, response, next) => {
-  Blog.findById(request.params.id)
-    .then(blog => {
-      if (blog) {
+blogsRouter.get('/:id', async (request, response, next) => {
+  const blog = await Blog.findById(request.params.id)
+  try{    
+       if (blog) {
         response.json(blog)
       } else {
         response.status(404).end()
       }
-    })
-    .catch(error => next(error))
+     } catch(exception) {
+      next(exception)
+     }
 })
 
 blogsRouter.post('/', async (request, response, next) => {
   const body = request.body
+
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  const user = await User.findById(decodedToken.id)
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
   if (!body.likes) {
     body.likes = 0
 }
@@ -27,10 +38,13 @@ const blog = new Blog({
   title: body.title,
   author: body.author,
   url: body.url,
-  likes: body.likes
+  likes: body.likes,
+  user: user.id
 })
 try {
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
   response.status(201).json(savedBlog)
 } catch(exception) {
   next(exception)
@@ -38,25 +52,39 @@ try {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  try{
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
-} catch (exception) {
-  next(exception)
+  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+  const user = await User.findById(decodedToken.id)
+  const blogToDelete = await Blog.findById(request.params.id)
+
+  if ( blogToDelete.user._id.toString() === user._id.toString() ) {
+    try {
+        await Blog.findByIdAndDelete(request.params.id)
+        response.status(204).end()
+      } catch (exception) {
+        next(exception)
+      }
+} else {
+    return response.status(401).json({ error: `Unauthorized` })
 }
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
   const body = request.body
+  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+  const user = await User.findById(decodedToken.id)
+
   if (!body.likes) {
       body.likes = 0
   }
-  const blog = {
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes,
-    comments: body.comments,
+  const blogToUpdate = await Blog.findById(request.params.id)
+
+  if ( blogToUpdate.user._id.toString() === user._id.toString() ) {
+    const blog = {
+        title: body.title,
+        author: body.author,
+        url: body.url,
+        likes: body.likes
+    }
   }
   try{    
     const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
